REEB-287 fix image server bug
git-svn-id: http://svn.uhuila.cn/reeb/trunk@1046 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ cdn/image-server/app/controllers/Images.java
@@ -37,7 +37,7 @@ public class Images extends Controller {
      * @param thirdDir
      * @param imageName
      */
-    public static void showImage(int firstDir, int secondDir, int thirdDir, String imageName) {
+    public static void showImage(long firstDir, long secondDir, long thirdDir, String imageName) {
         int width = 0;
         int height = 0;
         String imageSizeType = "";
